Update API/Controllers/RepliesController.cs Update API/Controllers/TicketsController.cs Update BusinessLogic/Service/Application/ReplyService.cs Update BusinessLogic/Service/Application/TicketService.cs Update Common/Repository/Application/ReplyRepository.cs Update Common/Repository/Application/TicketRepository.cs Update DataAccess/Models/Ticket.cs

+++ BusinessLogic/Service/Application/TicketService.cs
@@ -23,7 +23,15 @@ namespace BusinessLogic.Service.Application
 
         public bool Delete(int id)
         {
-            return _iTicketRepository.Delete(id);
+            if (string.IsNullOrWhiteSpace(id.ToString()))
+            {
+                return status;
+            }
+            else
+            {
+                return _iTicketRepository.Delete(id);
+            }
+            
         }
 
         public List<Ticket> Get()
@@ -55,7 +63,15 @@ namespace BusinessLogic.Service.Application
 
         public bool Update(int id, TicketVM ticketVM)
         {
-            return _iTicketRepository.Update(id, ticketVM);
+            if (string.IsNullOrWhiteSpace(ticketVM.Id.ToString()) && string.IsNullOrWhiteSpace(ticketVM.Status_Id.ToString()))
+            {
+                return status;
+            }
+            else
+            {
+                return _iTicketRepository.Update(id, ticketVM);
+            }
+            
         }
     }
 }
